pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "mcp[cli]>=1.8.0",
    "odoorpc>=0.10.1",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
]
